+++ app/Http/Livewire/Dashboard/Latest.php
@@ -8,17 +8,17 @@ use Livewire\Component;
 class Latest extends Component
 {
     /**
-     * @var mixed $latest
+     * @var mixed
      */
     public $latest;
 
     public function mount(): void
     {
-        $this->latest = Configuration::latest("updated_at")->take(5)->get();
+        $this->latest = Configuration::latest('updated_at')->take(5)->get();
     }
+
     public function render(): \Illuminate\View\View
     {
-
         return view('livewire.dashboard.latest');
     }
 }