+++ app/Providers/RouteServiceProvider.php
@@ -23,7 +23,6 @@ class RouteServiceProvider extends ServiceProvider
      * The controller namespace for the application.
      *
      * When present, controller route declarations will automatically be prefixed with this namespace.
-     *
      */
     // protected $namespace = 'App\\Http\\Controllers';