+++ app/Traits/Form/CodeQuality.php
@@ -5,17 +5,29 @@ namespace App\Traits\Form;
 trait CodeQuality
 {
     public bool $stepExecutePhpunit; //true
+
     public bool $stepExecutePestphp; //false
+
     public bool $stepSecurityCheck; //false
+
     public bool $stepExecuteCodeSniffer; //false
+
     public string $stepDirCodeSniffer; // app
+
     public bool $stepInstallCodeSniffer; //true
+
     public bool $stepExecuteStaticAnalysis; // false
+
     public string $stepDirStaticAnalysis; // app
+
     public string $stepToolStaticAnalysis; // phpstan|larastan|psalmlaravel
+
     public bool $stepInstallStaticAnalysis; //true
+
     public bool $stepDusk; // false
+
     public bool $stepPhpstanUseNeon; // false
+
     public bool $stepPsalmReport; // false
 
     public function loadDefaultsCodeQuality(): void
@@ -24,11 +36,11 @@ trait CodeQuality
         $this->stepExecutePestphp = false;
         $this->stepSecurityCheck = false;
         $this->stepExecuteCodeSniffer = false;
-        $this->stepDirCodeSniffer = "app";
+        $this->stepDirCodeSniffer = 'app';
         $this->stepInstallCodeSniffer = true;
         $this->stepExecuteStaticAnalysis = false;
-        $this->stepDirStaticAnalysis = "app";
-        $this->stepToolStaticAnalysis = "larastan";
+        $this->stepDirStaticAnalysis = 'app';
+        $this->stepToolStaticAnalysis = 'larastan';
         $this->stepInstallStaticAnalysis = true;
         $this->stepDusk = false;
         $this->stepPhpstanUseNeon = false;
@@ -37,15 +49,15 @@ trait CodeQuality
 
     public function loadCodeQualityFromJson(object $j): void
     {
-        data_fill($j, "stepDirCodeSniffer", "app");
-        data_fill($j, "stepInstallCodeSniffer", true);
-        data_fill($j, "stepDirStaticAnalysis", "app");
-        data_fill($j, "stepInstallStaticAnalysis", true);
-        data_fill($j, "stepToolStaticAnalysis", 'larastan');
-        data_fill($j, "stepExecutePestphp", false);
-        data_fill($j, "stepPhpstanUseNeon", false);
-        data_fill($j, "stepPsalmReport", false);
-        data_fill($j, "stepSecurityCheck", false);
+        data_fill($j, 'stepDirCodeSniffer', 'app');
+        data_fill($j, 'stepInstallCodeSniffer', true);
+        data_fill($j, 'stepDirStaticAnalysis', 'app');
+        data_fill($j, 'stepInstallStaticAnalysis', true);
+        data_fill($j, 'stepToolStaticAnalysis', 'larastan');
+        data_fill($j, 'stepExecutePestphp', false);
+        data_fill($j, 'stepPhpstanUseNeon', false);
+        data_fill($j, 'stepPsalmReport', false);
+        data_fill($j, 'stepSecurityCheck', false);
         $this->stepExecutePhpunit = $j->stepExecutePhpunit;
         $this->stepExecutePestphp = $j->stepExecutePestphp;
         $this->stepExecuteCodeSniffer = $j->stepExecuteCodeSniffer;
@@ -62,24 +74,24 @@ trait CodeQuality
     }
 
     /**
-     * @param array<mixed> $data
+     * @param  array<mixed>  $data
      * @return array<mixed>
      */
     public function setDataCodeQuality(array $data): array
     {
-        $data["stepExecutePhpunit"] = $this->stepExecutePhpunit;
-        $data["stepExecutePestphp"] = $this->stepExecutePestphp;
-        $data["stepExecuteCodeSniffer"] = $this->stepExecuteCodeSniffer;
-        $data["stepDirCodeSniffer"] = $this->stepDirCodeSniffer;
-        $data["stepInstallCodeSniffer"] = $this->stepInstallCodeSniffer;
-        $data["stepExecuteStaticAnalysis"] = $this->stepExecuteStaticAnalysis;
-        $data["stepDirStaticAnalysis"] = $this->stepDirStaticAnalysis;
-        $data["stepToolStaticAnalysis"] = $this->stepToolStaticAnalysis;
-        $data["stepInstallStaticAnalysis"] = $this->stepInstallStaticAnalysis;
-        $data["stepDusk"] = $this->stepDusk;
-        $data["stepPhpstanUseNeon"] = $this->stepPhpstanUseNeon;
-        $data["stepPsalmReport"] = $this->stepPsalmReport;
-        $data["stepSecurityCheck"] = $this->stepSecurityCheck;
+        $data['stepExecutePhpunit'] = $this->stepExecutePhpunit;
+        $data['stepExecutePestphp'] = $this->stepExecutePestphp;
+        $data['stepExecuteCodeSniffer'] = $this->stepExecuteCodeSniffer;
+        $data['stepDirCodeSniffer'] = $this->stepDirCodeSniffer;
+        $data['stepInstallCodeSniffer'] = $this->stepInstallCodeSniffer;
+        $data['stepExecuteStaticAnalysis'] = $this->stepExecuteStaticAnalysis;
+        $data['stepDirStaticAnalysis'] = $this->stepDirStaticAnalysis;
+        $data['stepToolStaticAnalysis'] = $this->stepToolStaticAnalysis;
+        $data['stepInstallStaticAnalysis'] = $this->stepInstallStaticAnalysis;
+        $data['stepDusk'] = $this->stepDusk;
+        $data['stepPhpstanUseNeon'] = $this->stepPhpstanUseNeon;
+        $data['stepPsalmReport'] = $this->stepPsalmReport;
+        $data['stepSecurityCheck'] = $this->stepSecurityCheck;
 
         return $data;
     }