+++ app/Traits/Form/Deploy.php
@@ -5,9 +5,13 @@ namespace App\Traits\Form;
 trait Deploy
 {
     public string $stepDeployType;
+
     public string $stepDeployWebhookType;
+
     public string $stepDeployWebhookUrl;
+
     public string $stepDeployForgeServerName;
+
     public string $stepDeployForgeSiteName;
     //public $stepDeployApiToken;
 
@@ -15,20 +19,20 @@ trait Deploy
     {
         $this->stepDeployType = 'none';
         $this->stepDeployWebhookType = 'secret';
-        $this->stepDeployWebhookUrl = "WEBHOOK_URL";
-        $this->stepDeployForgeServerName = "";
-        $this->stepDeployForgeSiteName = "";
+        $this->stepDeployWebhookUrl = 'WEBHOOK_URL';
+        $this->stepDeployForgeServerName = '';
+        $this->stepDeployForgeSiteName = '';
 
         //$this->stepDeployApiToken = "VAPOR_API_TOKEN";
     }
 
     public function loadDeployFromJson(object $j): void
     {
-        data_fill($j, "stepDeployType", 'none');
-        data_fill($j, "stepDeployWebhookType", 'secret');
-        data_fill($j, "stepDeployWebhookUrl", 'WEBHOOK_URL');
-        data_fill($j, "stepDeployForgeServerName", '');
-        data_fill($j, "stepDeployForgeSiteName", '');
+        data_fill($j, 'stepDeployType', 'none');
+        data_fill($j, 'stepDeployWebhookType', 'secret');
+        data_fill($j, 'stepDeployWebhookUrl', 'WEBHOOK_URL');
+        data_fill($j, 'stepDeployForgeServerName', '');
+        data_fill($j, 'stepDeployForgeSiteName', '');
         //data_fill($j, "stepDeployApiToken", 'VAPOR_API_TOKEN');
 
         $this->stepDeployType = $j->stepDeployType;
@@ -39,16 +43,16 @@ trait Deploy
     }
 
     /**
-     * @param array<mixed> $data
+     * @param  array<mixed>  $data
      * @return array<mixed>
      */
     public function setDeployData(array $data)
     {
-        $data["stepDeployType"] = $this->stepDeployType;
-        $data["stepDeployWebhookType"] = $this->stepDeployWebhookType;
-        $data["stepDeployWebhookUrl"] = $this->stepDeployWebhookUrl;
-        $data["stepDeployForgeSiteName"] = $this->stepDeployForgeSiteName;
-        $data["stepDeployForgeServerName"] = $this->stepDeployForgeServerName;
+        $data['stepDeployType'] = $this->stepDeployType;
+        $data['stepDeployWebhookType'] = $this->stepDeployWebhookType;
+        $data['stepDeployWebhookUrl'] = $this->stepDeployWebhookUrl;
+        $data['stepDeployForgeSiteName'] = $this->stepDeployForgeSiteName;
+        $data['stepDeployForgeServerName'] = $this->stepDeployForgeServerName;
         //$data["stepDeployApiToken"] = $this->stepDeployApiToken;
 
         return $data;