+++ tests/Feature/DatabaseFormTest.php
@@ -1,127 +1,73 @@
 <?php
 
-namespace Tests\Feature;
+test('database default')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test')
+    ->call('submitForm')
+    ->assertSee(readAsset('on-push-branches.yaml'))
+    ->assertSee(readAsset('mysql-service.yaml'));
 
-use App\Http\Livewire\ConfiguratorForm;
-use Illuminate\Foundation\Testing\DatabaseMigrations;
-use Illuminate\Foundation\Testing\RefreshDatabase;
-use Illuminate\Foundation\Testing\WithFaker;
-use Livewire\Livewire;
-use Tests\TestCase;
+test('database default hints for migration')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test')
+    ->set('stepRunMigrations', false)
+    ->call('submitForm')
+    ->assertSee(readAsset('on-push-branches.yaml'))
+    ->assertSee(readAsset('mysql-service.yaml'))
+    ->assertCount('hints', 1);
 
-class DatabaseFormTest extends TestCase
-{
-    use DatabaseMigrations;
+test('no database and no migrations')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test No Database')
+    ->set('onPullrequest', true)
+    ->set('databaseType', 'none')
+    ->set('stepRunMigrations', false)
+    ->call('submitForm')
+    ->assertHasNoErrors('yaml')
+    ->assertSet('hints', [])
+    ->assertDontSee('image: mysql:')
+    ->assertDontSee('DB_CONNECTION: sqlite');
 
-    const DIR_MOCK ="tests/Feature/mock-asserts/";
+test('generates hint - no database and migrations')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test No Database')
+    ->set('onPullrequest', true)
+    ->set('databaseType', 'none')
+    ->call('submitForm')
+    ->assertHasNoErrors('yaml')
+    ->assertCount('hints', 1)
+    ->assertDontSee('image: mysql:');
 
+test('with sqlite database and migrations')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test No Database')
+    ->set('onPullrequest', true)
+    ->set('databaseType', 'sqlite')
+    ->call('submitForm')
+    ->assertHasNoErrors('yaml')
+    ->assertCount('hints', 0)
+    ->assertSee(readAsset('sqlite-migration.yaml'))
+    ->assertDontSee('image: mysql:');
 
-    public function test_databasedefault(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test")
-            ->call('submitForm')
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."on-push-branches.yaml")))
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."mysql-service.yaml")));
-    }
+test('test postgresql with migration')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test Postgresql')
+    ->set('onPullrequest', true)
+    ->set('databaseType', 'postgresql')
+    ->call('submitForm')
+    ->assertHasNoErrors('yaml')
+    ->assertCount('hints', 0)
+    ->assertSee(readAsset('postgresql-service.yaml'))
+    ->assertDontSee('image: mysql:');
 
-
-    public function test_databasedefault_hintsformigration(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test")
-            ->set("stepRunMigrations", false)
-            ->call('submitForm')
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."on-push-branches.yaml")))
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."mysql-service.yaml")))
-            ->assertCount("hints", 1);
-    }
-
-    /**
-     * Testing with no database and no migrations.
-     * @return void
-     */
-    public function test_nodatabase(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test No Database")
-            ->set("onPullrequest", true)
-            ->set("databaseType", "none")
-            ->set("stepRunMigrations", false)
-            ->call('submitForm')
-            ->assertHasNoErrors('yaml')
-            ->assertSet('hints', [])
-            ->assertDontSee("image: mysql:")
-            ->assertDontSee("DB_CONNECTION: sqlite");
-    }
-
-    /**
-     * Testing with no database and migrations, it generates hint.
-     * @return void
-     */
-    public function test_nodatabase_withmigration(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test No Database")
-            ->set("onPullrequest", true)
-            ->set("databaseType", "none")
-            ->call('submitForm')
-            ->assertHasNoErrors('yaml')
-            ->assertCount('hints', 1)
-            ->assertDontSee("image: mysql:");
-    }
-
-    /**
-     * Testing with sqlite database and migrations
-     * @return void
-     */
-    public function test_sqlite_withmigration(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test No Database")
-            ->set("onPullrequest", true)
-            ->set("databaseType", "sqlite")
-            ->call('submitForm')
-            ->assertHasNoErrors('yaml')
-            ->assertCount('hints', 0)
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."sqlite-migration.yaml")))
-            ->assertDontSee("image: mysql:");
-    }
-
-    /**
-     * Testing with sqlite database and migrations
-     * @return void
-     */
-    public function test_postgresql_withmigration(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test Postgresql")
-            ->set("onPullrequest", true)
-            ->set("databaseType", "postgresql")
-            ->call('submitForm')
-            ->assertHasNoErrors('yaml')
-            ->assertCount('hints', 0)
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."postgresql-service.yaml")))
-            ->assertDontSee("image: mysql:");
-    }
-
-    /**
-     * Testing with sqlite database and migrations
-     * @return void
-     */
-    public function test_postgresql_withnomigration(): void
-    {
-        Livewire::test(ConfiguratorForm::class)
-            ->set("name","Test Postgresql")
-            ->set("onPullrequest", true)
-            ->set("databaseType", "postgresql")
-            ->set("stepRunMigrations", false)
-            ->call('submitForm')
-            ->assertHasNoErrors('yaml')
-            ->assertCount('hints', 1)
-            ->assertSee(file_get_contents(base_path(self::DIR_MOCK."postgresql-service.yaml")))
-            ->assertDontSee("image: mysql:");
-    }
-
-
-}
+test('postgresql with no migration')
+    ->livewire(ConfiguratorForm())
+    ->set('name', 'Test Postgresql')
+    ->set('onPullrequest', true)
+    ->set('databaseType', 'postgresql')
+    ->set('stepRunMigrations', false)
+    ->call('submitForm')
+    ->assertHasNoErrors('yaml')
+    ->assertCount('hints', 1)
+    ->assertSee(readAsset('postgresql-service.yaml'))
+    ->assertDontSee('image: mysql:');